♻️ refactor(aiAgent): extract toolDiscovery and heteroDispatch pipeline stages
Part of LOBE-13706 (PR3 of 4). Mechanical extraction of the two largest
execAgent stages into explicit pipeline functions:

- pipeline/heteroDispatch.ts: the stage-3.5 heterogeneous-agent early exit
  (device gateway / cloud sandbox handoff, ~745 lines)
- pipeline/toolDiscovery.ts: stages 5/5a-5f tool discovery (plugins,
  connectors, skills, composio, device pool + execution plan, tools engine,
  activator manifest map, ~945 lines)
- types.ts: new ExecRunContext shared by pipeline stages, replacing implicit
  closure-variable passing

index.ts shrinks from 5673 to 4124 lines. aiAgentService exports and
behavior are unchanged; existing __tests__ pass unmodified.
