+++ apps/server/src/services/aiAgent/types.ts
@@ -9,9 +9,46 @@ import type {
   WorkspaceInitResult,
 } from '@lobechat/types';
 
+import type { AgentConfigWithId } from '@/server/services/agent';
+
 import type { EvalContext } from '@/server/modules/Mecha/ContextEngineering/types';
 import type { AgentHook } from '@/server/services/agentRuntime/hooks/types';
 import type { EvalRuntimeContext } from '@/server/services/agentRuntime/types';
+import type { DeviceAccessReason } from './deviceAccessPolicy';
+
+/**
+ * Resolved run state shared by the {@link AiAgentService.execAgent} pipeline
+ * stages (`pipeline/*`). Built once inside `execAgent` after agent/topic/turn
+ * setup, then handed to each extracted stage so the data every stage consumes
+ * is explicit instead of riding on closure variables.
+ *
+ * `agentConfig` is intentionally the same MUTABLE object `execAgent` holds:
+ * stages append to `systemRole` (connector ownership notes, project
+ * instructions) and later steps — `createOperation` in particular — must see
+ * those writes.
+ */
+export interface ExecRunContext {
+  agentConfig: AgentConfigWithId;
+  appContext?: InternalExecAgentParams['appContext'];
+  /** Persisted assistant placeholder row id (spinner anchor / error sink). */
+  assistantMessageId: string;
+  canUseDevice: boolean;
+  deviceAccessReason: DeviceAccessReason;
+  /** Effective model for this run (topic-pinned model already applied). */
+  model: string;
+  parentMessageId?: string;
+  /** Persistence-attribution agent id (Agent Signal marker aware). */
+  persistAgentId: string;
+  prompt: string;
+  provider: string;
+  /** The actual executing agent row id resolved from id/slug. */
+  resolvedAgentId: string;
+  /** Topic id — guaranteed to exist by the time pipeline stages run. */
+  topicId: string;
+  trigger?: string;
+  /** User turn row id; undefined when the run starts from history (resume). */
+  userMessageId?: string;
+}
 
 /**
  * Internal params for execAgent with step lifecycle callbacks
